split GetNewBlockV3 IT to avoid assumeThat spreading around the code
(cherry picked from commit 149ce7b8186119c745246730a25fb485374f47ed)
